requirements.txt
asgiref==3.9.1
Django==5.2.5
django-cors-headers==4.7.0
djangorestframework==3.16.1
packaging==25.0
pillow==11.3.0
pytesseract==0.3.13
python-decouple==3.8
sqlparse==0.5.3
tzdata==2025.2
opencv-python==4.12.0.88
